requirements.txt
wheel~=0.36.2
setuptools~=57.0.0
filetype~=1.2.0
PyYAML~=6.0
Flask~=2.0.3
MarkupSafe~=2.0.1
Werkzeug~=2.0.3
Jinja2~=3.0.3
click~=8.0.4
itsdangerous~=2.0.1
bottle~=0.12.25
pywebview~=4.0.2
pycparser~=2.21
zipp~=3.6.0
dataclasses~=0.6
pywin32~=305